use zustand to store the loglines
